requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
pgvector==0.2.4
python-dotenv==1.0.0
openai==1.3.7
langchain==0.1.0
langchain-openai==0.0.2
pypdf==3.17.4
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
pandas==2.1.4
redis==5.0.1
tiktoken==0.5.2
numpy==1.26.2
sentence-transformers==2.2.2
